requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
groq==0.4.2
together==0.2.8
deepseek==1.0.0
google-generativeai==0.3.2
openai==1.12.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
aiofiles==23.2.1
python-magic==0.4.27
python-docx==1.0.1 